fix: unify command dispatch across transports and gate telephony on relay path
The WebSocket relay (RelayClient) and local HTTP server (BridgeRouter) each
carried a full copy of the command surface, and they had drifted:

  - DeviceCapabilities.hasTelephony gating existed only in BridgeRouter, so
    /send_sms, /call, /contacts arriving over the relay (the path Hermes
    actually drives the device through) were ungated on AAOS devices.
  - /open_app read `packageName` in BridgeRouter but the client sends
    `package` (android_tool.py), so HTTP /open_app could never deserialize.

Extract a single CommandDispatcher.dispatch(method, path, params, body,
authenticated) as the source of truth, with telephony gating applied. Both
transports become thin adapters that parse a request and call it. Relay logic
is preserved byte-for-byte except the three gated endpoints; BridgeRouter now
inherits the proven relay behaviour (and the /open_app fix) for free.
